+++ betterrandom/src/test/java/io/github/pr0methean/betterrandom/seed/RandomDotOrgUtils.java
@@ -5,7 +5,12 @@ import io.github.pr0methean.betterrandom.TestingDeficiency;
 import java.net.InetSocketAddress;
 import java.net.Proxy;
 import java.net.Proxy.Type;
+import java.security.KeyManagementException;
+import java.security.NoSuchAlgorithmException;
 import java.util.UUID;
+import javax.net.ssl.SSLContext;
+import javax.net.ssl.SSLSocketFactory;
+import org.slf4j.LoggerFactory;
 
 /**
  * Utility methods used in {@link RandomDotOrgSeedGeneratorHermeticTest} and
@@ -41,4 +46,19 @@ public enum RandomDotOrgUtils {
     return !TestUtils.isAppveyor() && !("osx".equals(System.getenv("TRAVIS_OS_NAME")));
   }
 
-}
+  public static SSLSocketFactory createSocketFactory() {
+    try {
+      String[] versionNumberParts = System.getProperty("java.version").split("\\.");
+      int javaVersion = Integer.parseInt(versionNumberParts[0]);
+      if (javaVersion == 1) {
+        javaVersion = Integer.parseInt(versionNumberParts[1]);
+      }
+      LoggerFactory.getLogger(RandomDotOrgUtils.class)
+          .info("Detected JDK version {0}", javaVersion);
+      SSLContext context = SSLContext.getInstance(javaVersion >= 11 ? "TLSv1.3" : "TLSv1.2");
+      context.init(null, null, null);
+      return context.getSocketFactory();
+    } catch (NoSuchAlgorithmException | KeyManagementException e) {
+      throw new AssertionError(e);
+    }
+  }}